Renamed backend.ts to helpers.ts
