+++ test/createOption.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 import { OptionState } from "../src/utils/types";
-import { buyer, seller, initializer, dummyOptionRequest, publishDummyOptionRequest } from "./utils/backend";
+import { buyer, seller, initializer, dummyOptionRequest, publishDummyOptionRequest } from "./utils/helpers";
 import { NFTOptContract, deployMainContract } from "../src/utils/deployment";
 
 describe("createOption", function () {
